Struct: Reorganize the main module (lime_python) in 3 sub modules (base, documents, utils)
