+++ lime_python/base/__init__.py
@@ -1,0 +1,8 @@
+from lime_python.base.mediaType import MediaType
+from lime_python.base.document import Document
+from lime_python.base.identity import Identity
+from lime_python.base.node import Node
+from lime_python.base.envelope import Envelope
+from lime_python.base.message import Message
+from lime_python.base.notification import Notification, NotificationEvent
+from lime_python.base.command import Command, CommandMethod, CommandStatus
