+++ lime_python/base/message.py
@@ -1,6 +1,6 @@
-from lime_python.mediaType import MediaType
-from lime_python.envelope import Envelope
-from lime_python.document import Document
+from lime_python.base.mediaType import MediaType
+from lime_python.base.envelope import Envelope
+from lime_python.base.document import Document
 
 
 class Message(Envelope):
