+++ lime_python/documents/__init__.py
@@ -1,31 +1,18 @@
-from lime_python.utils.stringUtils import StringUtils
-from lime_python.base.document import Document
-from lime_python.documents.chatStateDocument import ChatStateDocument, \
+from lime_python.documents.chatStateDocument import ChatStateDocument,\
     ChatState
 from lime_python.documents.collectionDocument import CollectionDocument
-from lime_python.base.command import Command, CommandMethod, CommandStatus
 from lime_python.documents.containerDocument import ContainerDocument
-from lime_python.base.mediaType import MediaType
-from lime_python.base.envelope import Envelope
-from lime_python.utils.header import Header
-from lime_python.base.identity import Identity
 from lime_python.documents.inputDocument import InputDocument, Validation, Rule
 from lime_python.documents.listDocument import ListDocument
 from lime_python.documents.locationDocument import LocationDocument
 from lime_python.documents.mediaLinkDocument import MediaLinkDocument
 from lime_python.documents.menuDocument import MenuDocument
-from lime_python.base.message import Message
 from lime_python.documents.multimediaMenuDocument import MultimediaMenuDocument
-from lime_python.base.node import Node
-from lime_python.base.notification import Notification, NotificationEvent
 from lime_python.documents.paymentInvoiceDocument import PaymentInvoiceDocument
 from lime_python.documents.paymentReceiptDocument import PaymentReceiptDocument
-from lime_python.documents.plainTextDocument import PlainTextDocument
-from lime_python.utils.reason import Reason
-from lime_python.utils.reasonCode import ReasonCode
 from lime_python.documents.redirectDocument import RedirectDocument
+from lime_python.documents.plainTextDocument import PlainTextDocument
 from lime_python.documents.resourceDocument import ResourceDocument
-from lime_python.utils.scope import Scope
 from lime_python.documents.sensitiveInformationDocument \
     import SensitiveInformationDocument
 from lime_python.documents.webLinkDocument import WebLinkDocument, Target
