+++ lime_python/documents/inputDocument.py
@@ -1,6 +1,6 @@
-from lime_python.plainTextDocument import PlainTextDocument
-from lime_python.mediaType import MediaType as MT
-from lime_python.document import Document
+from lime_python.documents.plainTextDocument import PlainTextDocument
+from lime_python.base.mediaType import MediaType as MT
+from lime_python.base.document import Document
 from enum import Enum
