+++ lime_python/documents/locationDocument.py
@@ -1,5 +1,5 @@
-from lime_python.mediaType import MediaType
-from lime_python.document import Document
+from lime_python.base.mediaType import MediaType
+from lime_python.base.document import Document
 
 
 class _LocationDocument(Document):
