+++ lime_python/documents/mediaLinkDocument.py
@@ -1,5 +1,5 @@
-from lime_python.mediaType import MediaType as MT
-from lime_python.document import Document
+from lime_python.base.mediaType import MediaType as MT
+from lime_python.base.document import Document
 
 
 class _MediaLinkDocument(Document):
