+++ lime_python/documents/multimediaMenuDocument.py
@@ -1,15 +1,14 @@
-from lime_python.mediaType import MediaType
-from lime_python.document import Document
-from lime_python.header import Header
-from lime_python.scope import Scope
+from lime_python.base.mediaType import MediaType
+from lime_python.base.document import Document
+from lime_python.utils.header import Header
+from lime_python.utils.scope import Scope
 
 
-class _MenuDocument(Document):
+class _MultimediaMenuDocument(Document):
 
-    MIME_TYPE = 'application/vnd.lime.select+json'
+    MIME_TYPE = 'application/vnd.lime.document-select+json'
 
     def __init__(self, scope=Scope.Transient, header=None, options=[]):
-        super().__init__(MediaType.Parse(_MenuDocument.MIME_TYPE))
 
         self.Scope = scope
         self.Header = header
@@ -168,22 +167,18 @@ class _MenuDocument(Document):
                         'value': self.GetLabelDocumentJson()
                     }
                 })
-            if isinstance(self.Value, Document):
+            if isinstance(self.Value, Document) \
+                    or isinstance(self.Value, dict):
                 json.update({
                     'value': {
                         'type': str(self.GetValueMediaType()),
                         'value': self.GetValueDocumentJson()
                     }
                 })
-            elif isinstance(self.Value, dict):
-                json.update({
-                    'type': str(self.GetValueMediaType()),
-                    'value': self.GetValueDocumentJson()
-                })
 
             return json
 
 
-class MenuDocument(_MenuDocument):
+class MultimediaMenuDocument(_MultimediaMenuDocument):
 
-    Type = MediaType.Parse(_MenuDocument.MIME_TYPE)
+    Type = MediaType.Parse(_MultimediaMenuDocument.MIME_TYPE)
