+++ lime_python/utils/header.py
@@ -1,4 +1,4 @@
-from lime_python.document import Document
+from lime_python.base.document import Document
 
 
 class Header:
