+++ lime_python/utils/reason.py
@@ -1,4 +1,4 @@
-from lime_python.reasonCode import ReasonCode
+from lime_python.utils.reasonCode import ReasonCode
 
 
 class Reason:
