+++ lime_python/utils/scope.py
