+++ setup.py
@@ -4,18 +4,18 @@ from distutils.core import setup
 setup(
     name='lime_python',
     packages=['lime_python'],
-    version='0.3',
+    version='0.5',
     license='apache-2.0',
     description='This module implements the basic models of the LIME protocol',
     author='Gabriel Rodrigues dos Santos',
     author_email='gabrielr@take.net',
     url='https://github.com/takenet/lime-python',
-    download_url='https://github.com/takenet/lime-python/archive/v_03.tar.gz',
+    download_url='https://github.com/takenet/lime-python/archive/v_05.tar.gz',
     keywords=['lime', 'blip', 'chatbot'],
     install_requires=[],
     classifiers=[
         # "3 - Alpha", "4 - Beta" or "5 - Production/Stable"
-        'Development Status :: 3 - Alpha',
+        'Development Status :: 4 - Beta',
         'Intended Audience :: Developers',
         'Topic :: Software Development :: Build Tools',
         'License :: OSI Approved :: Apache Software License',
